refactor: move graph layout out of select/ into graph/layout
`graph_layout.rs` converts a `ChangeGraph` into a jj-log-style row list
and has no ratatui/crossterm dependencies — it is pure layout, not TUI.
Move it from `src/select/` to `src/graph/layout.rs` so upcoming non-TUI
consumers (a revamped `stakk show` graph renderer) can reuse it without
reaching into the TUI module.

No behavior change:

- module moved verbatim; imports in `select/{app,graph_widget,bookmark_widget}`
  now use `crate::graph::layout`
- insta snapshot moved to `src/graph/snapshots/` (filename encodes the
  module path) with its `source:` metadata updated
