+++ CLAUDE.md
@@ -91,10 +91,10 @@ src/
 │   ├── comment.rs   # Stack comment formatting, parsing, and template context
 │   └── default_comment.md.jinja  # Default minijinja template for stack comments
 ├── graph/           # Change graph construction (ChangeGraph, BookmarkSegment, BranchStack)
+│   └── layout.rs    # Convert ChangeGraph → jj-log-style row list (commit tree + display rows)
 ├── select/          # Interactive TUI selection (ratatui inline viewport)
 │   ├── mod.rs       # Public API: resolve_bookmark_interactively(), SelectionResult
 │   ├── app.rs       # App state machine, event loop, terminal init
-│   ├── graph_layout.rs  # Convert ChangeGraph → jj-log-style row list (commit tree + display rows)
 │   ├── graph_widget.rs  # Screen 1: jj-log-style graph widget (leaf selection, collapsing, scrolling)
 │   ├── bookmark_widget.rs # Screen 2: bookmark toggle/assignment widget
 │   ├── bookmark_gen.rs # Bookmark validation and external command execution
